# apis/services/hr-service/docs/swagger.yaml
basePath: /
definitions:
  erp_hr-service_internal_models.Employee:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      department:
        type: string
      email:
        type: string
      employee_id:
        type: string
      id:
        type: integer
      manager:
        $ref: '#/definitions/erp_hr-service_internal_models.Employee'
      manager_id:
        type: integer
      name:
        type: string
      position:
        type: string
      role:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the HR microservice for the ERP system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HR Service API
  version: "1.0"
paths:
  /api/v1/hr/attendance/stats:
    get:
      consumes:
      - application/json
      description: Retrieve summary statistics for today's attendance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get attendance statistics
      tags:
      - hr-stubs
  /api/v1/hr/employees:
    get:
      consumes:
      - application/json
      description: Get a list of all employees with their managers preloaded
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/erp_hr-service_internal_models.Employee'
            type: array
      summary: Get all employees
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Create a new employee and log the audit trail
      parameters:
      - description: Employee to create
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/erp_hr-service_internal_models.Employee'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/erp_hr-service_internal_models.Employee'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new employee
      tags:
      - employees
  /api/v1/hr/employees/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about an employee by their ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erp_hr-service_internal_models.Employee'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an employee by ID
      tags:
      - employees
    patch:
      consumes:
      - application/json
      description: Update employee information and log status changes
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated employee data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/erp_hr-service_internal_models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erp_hr-service_internal_models.Employee'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing employee
      tags:
      - employees
  /api/v1/hr/leaves/pending:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all leave requests that are currently pending
        approval
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get pending leave requests
      tags:
      - hr-stubs
  /api/v1/hr/payroll/pending:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all payroll entries that are currently pending
        processing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get pending payroll entries
      tags:
      - hr-stubs
swagger: "2.0"